+++ Sources/StencilSwiftKit/Environment.swift
@@ -4,6 +4,7 @@
 // MIT Licence
 //
 
+import PathKit
 import Stencil
 
 public extension Extension {
@@ -59,10 +60,14 @@ private extension Extension {
 
 /// Creates an Environment for Stencil to load & render templates
 ///
+/// - Parameters:
+///   - templatePaths: Paths where Stencil can search for templates, used for example for `include` tags
+///   - extensions: Additional extensions with filters/tags/… you want to provide to Stencil
 /// - Returns: A fully configured `Environment`
-public func stencilSwiftEnvironment() -> Environment {
+public func stencilSwiftEnvironment(templatePaths: [Path] = [], extensions: [Extension] = []) -> Environment {
+  let loader = FileSystemLoader(paths: templatePaths)
   let ext = Extension()
   ext.registerStencilSwiftExtensions()
 
-  return Environment(extensions: [ext], templateClass: StencilSwiftTemplate.self)
+  return Environment(loader: loader, extensions: extensions + [ext], templateClass: StencilSwiftTemplate.self)
 }